Revert back from Text.Json to Newtonsoft.Json
This is a manual reversion of the commit cb0a2d7d.
Auto-reverting doesn't work due to the
refactoring after the commit.

System.Text.Json follows a strict mode while Newtonsoft.Json follows a
flexible mode by defaut. I deleted the unittest to ensure the strict
deserialization of integer types, since it's not a trivial job to
implement all the strict checks in Newtonsoft.Json. There is no
setting option provided in most cases. Let's stay with the default
behavior of Newtonsoft.Json.